refactor: extract android details helpers
