+++ src/managers/android/mod.rs
@@ -277,6 +277,7 @@
 //! This ensures consistent, predictable device ordering without hardcoded device lists.
 //!
 
+mod details;
 mod parser;
 mod sdk;
 mod version;
@@ -914,333 +915,6 @@ impl AndroidManager {
         } else {
             Ok(all_devices)
         }
-    }
-
-    /// Get the AVD directory path for a given AVD name
-    async fn get_avd_path(&self, avd_name: &str) -> Result<Option<PathBuf>> {
-        let avd_output = self
-            .command_executor
-            .run(&self.avdmanager_path, &["list", "avd"])
-            .await
-            .context("Failed to list Android AVDs")?;
-
-        let mut current_name = String::new();
-
-        for line in avd_output.lines() {
-            let trimmed = line.trim();
-            if let Some(caps) = AVD_NAME_REGEX.captures(trimmed) {
-                current_name = caps[1].to_string();
-            } else if let Some(caps) = PATH_REGEX.captures(trimmed) {
-                if current_name == avd_name {
-                    return Ok(Some(PathBuf::from(caps[1].to_string())));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-
-    /// Fine-tune AVD configuration after creation with avdmanager
-    async fn fine_tune_avd_config(
-        &self,
-        avd_name: &str,
-        config: &DeviceConfig,
-        _tag: &str,
-        _abi: &str,
-    ) -> Result<()> {
-        if let Some(avd_path) = self.get_avd_path(avd_name).await? {
-            let config_path = avd_path.join(files::CONFIG_FILE);
-
-            // Read existing config created by avdmanager
-            let mut config_content = fs::read_to_string(&config_path)
-                .await
-                .context("Failed to read existing AVD configuration")?;
-
-            // Use the user's original name instead of generating from device type
-            let device_display_name = &config.name;
-
-            // Only override specific settings if needed
-            let ram_mb = if let Some(ram) = &config.ram_size {
-                ram.parse::<u32>().unwrap_or(0)
-            } else {
-                0
-            };
-
-            let storage_mb = if let Some(storage) = &config.storage_size {
-                storage.parse::<u32>().unwrap_or(0)
-            } else {
-                0
-            };
-
-            // Create AvdId format: Pixel_9_Pro_Fold
-            let avd_id = device_display_name.replace(" ", "_");
-
-            // Add or update avd.ini.displayname (with spaces)
-            if !device_display_name.is_empty() {
-                if config_content.contains("avd.ini.displayname=") {
-                    // Update existing line
-                    if let Some(start) = config_content.find("avd.ini.displayname=") {
-                        if let Some(end) = config_content[start..].find('\n') {
-                            let line_end = start + end;
-                            config_content.replace_range(
-                                start..line_end,
-                                &format!("avd.ini.displayname={device_display_name}"),
-                            );
-                        }
-                    }
-                } else {
-                    // Add new line after encoding line
-                    if let Some(encoding_pos) = config_content.find("avd.ini.encoding=UTF-8\n") {
-                        let insert_pos = encoding_pos + "avd.ini.encoding=UTF-8\n".len();
-                        config_content.insert_str(
-                            insert_pos,
-                            &format!("avd.ini.displayname={device_display_name}\n"),
-                        );
-                    } else {
-                        // Add at the beginning
-                        config_content = format!(
-                            "avd.ini.displayname={device_display_name}\navd.ini.encoding=UTF-8\n{config_content}"
-                        );
-                    }
-                }
-            }
-
-            // Add or update AvdId (with underscores)
-            if !avd_id.is_empty() {
-                if config_content.contains("AvdId=") {
-                    // Update existing line
-                    if let Some(start) = config_content.find("AvdId=") {
-                        if let Some(end) = config_content[start..].find('\n') {
-                            let line_end = start + end;
-                            config_content
-                                .replace_range(start..line_end, &format!("AvdId={avd_id}"));
-                        }
-                    }
-                } else {
-                    // Add new line after displayname
-                    if let Some(displayname_pos) = config_content.find("avd.ini.displayname=") {
-                        if let Some(line_end) = config_content[displayname_pos..].find('\n') {
-                            let insert_pos = displayname_pos + line_end + 1;
-                            config_content.insert_str(insert_pos, &format!("AvdId={avd_id}\n"));
-                        }
-                    }
-                }
-            }
-
-            // Update RAM size if specified
-            if ram_mb > 0 {
-                if let Some(start) = config_content.find("hw.ramSize=") {
-                    if let Some(end) = config_content[start..].find('\n') {
-                        let line_end = start + end;
-                        config_content
-                            .replace_range(start..line_end, &format!("hw.ramSize={ram_mb}"));
-                    }
-                }
-            }
-
-            // Update storage size if specified
-            if storage_mb > 0 {
-                if let Some(start) = config_content.find("disk.dataPartition.size=") {
-                    if let Some(end) = config_content[start..].find('\n') {
-                        let line_end = start + end;
-                        config_content.replace_range(
-                            start..line_end,
-                            &format!(
-                                "disk.dataPartition.size={}G",
-                                storage_mb / STORAGE_MB_TO_GB_DIVISOR
-                            ),
-                        );
-                    }
-                }
-            }
-
-            // Ensure image.sysdir.1 has trailing slash for consistency
-            if config_content.contains("image.sysdir.1=")
-                && !config_content.contains("image.sysdir.1=system-images/android-")
-            {
-                // This should not happen, but add safety check
-            } else if let Some(start) = config_content.find("image.sysdir.1=") {
-                if let Some(end) = config_content[start..].find('\n') {
-                    let line = &config_content[start..start + end];
-                    if !line.ends_with('/') {
-                        let line_end = start + end;
-                        config_content.replace_range(start..line_end, &format!("{line}/"));
-                    }
-                }
-            }
-
-            // Write the updated configuration
-            fs::write(&config_path, config_content)
-                .await
-                .context("Failed to write updated AVD configuration")?;
-        }
-        Ok(())
-    }
-
-    /// Get detailed information for a specific AVD
-    pub async fn get_device_details(
-        &self,
-        avd_name: &str,
-        cached_info: Option<(String, u32, String)>, // (device_type, api_level, android_version_name)
-    ) -> Result<crate::models::DeviceDetails> {
-        log::debug!("Getting device details for AVD: '{avd_name}'");
-
-        // Initialize details with cached values or defaults
-        let mut details = crate::models::DeviceDetails {
-            name: avd_name.to_string(),
-            status: "Unknown".to_string(),
-            platform: crate::models::Platform::Android,
-            device_type: cached_info
-                .as_ref()
-                .map(|(dt, _, _)| dt.clone())
-                .unwrap_or_default(),
-            api_level_or_version: cached_info
-                .as_ref()
-                .map(|(_, api, version)| format!("API {api} (Android {version})"))
-                .unwrap_or_default(),
-            ram_size: None,
-            storage_size: None,
-            resolution: None,
-            dpi: None,
-            device_path: None,
-            system_image: None,
-            identifier: avd_name.to_string(),
-        };
-
-        // Check if device is running
-        let running_avds = self.get_running_avd_names().await?;
-        let is_running = running_avds.contains_key(avd_name);
-        details.status = if is_running {
-            "Running".to_string()
-        } else {
-            "Stopped".to_string()
-        };
-
-        // Read config.ini for additional details
-        if let Ok(home_dir) = std::env::var(HOME) {
-            let config_path = std::path::PathBuf::from(&home_dir)
-                .join(files::android::AVD_DIR)
-                .join("avd")
-                .join(format!("{avd_name}.avd"))
-                .join(files::CONFIG_FILE);
-
-            log::debug!("Checking config path: {config_path:?}");
-            if config_path.exists() {
-                log::debug!("Config file found, reading details for {avd_name}");
-                if let Ok(config_content) = tokio::fs::read_to_string(&config_path).await {
-                    // Parse configuration values
-                    let mut api_level = 0u32;
-                    for line in config_content.lines() {
-                        if let Some((key, value)) = line.split_once('=') {
-                            match key.trim() {
-                                "hw.ramSize" => {
-                                    if let Ok(ram_mb) = value.trim().parse::<u64>() {
-                                        details.ram_size = Some(format!("{ram_mb} MB"));
-                                    }
-                                }
-                                "disk.dataPartition.size" => {
-                                    // Parse storage size (e.g., "8192M", "4G")
-                                    let value = value.trim();
-                                    if let Some(size_str) = value.strip_suffix('M') {
-                                        if let Ok(size_mb) = size_str.parse::<u64>() {
-                                            details.storage_size = Some(format!("{size_mb} MB"));
-                                        }
-                                    } else if let Some(size_str) = value.strip_suffix('G') {
-                                        if let Ok(size_gb) = size_str.parse::<u64>() {
-                                            details.storage_size = Some(format!(
-                                                "{} MB",
-                                                size_gb * STORAGE_MB_TO_GB_DIVISOR as u64
-                                            ));
-                                        }
-                                    }
-                                }
-                                "hw.lcd.width" => {
-                                    if let Ok(width) = value.trim().parse::<u32>() {
-                                        // Need to also get height to form resolution
-                                        details.resolution = Some(format!("{width}x?"));
-                                    }
-                                }
-                                "hw.lcd.height" => {
-                                    if let Ok(height) = value.trim().parse::<u32>() {
-                                        // Combine with width if available
-                                        if let Some(ref res) = details.resolution {
-                                            if res.contains("x?") {
-                                                let width = res.replace("x?", "");
-                                                details.resolution =
-                                                    Some(format!("{width}x{height}"));
-                                            }
-                                        } else {
-                                            details.resolution = Some(format!("?x{height}"));
-                                        }
-                                    }
-                                }
-                                "hw.lcd.density" => {
-                                    details.dpi = Some(format!("{dpi} DPI", dpi = value.trim()));
-                                }
-                                "image.sysdir.1" => {
-                                    details.system_image = Some(value.trim().to_string());
-                                    // Extract API level from image path
-                                    if let Some(caps) = IMAGE_SYSDIR_REGEX.captures(value.trim()) {
-                                        if let Ok(parsed_api) = caps[1].parse::<u32>() {
-                                            api_level = parsed_api;
-                                        }
-                                    }
-                                }
-                                "hw.device.name" => {
-                                    details.device_type = value.trim().to_string();
-                                }
-                                _ => {}
-                            }
-                        }
-                    }
-
-                    // Set API level and version
-                    if api_level > 0 {
-                        let version_name = self.get_android_version_name(api_level);
-                        details.api_level_or_version =
-                            format!("API {api_level} (Android {version_name})");
-                    }
-                }
-
-                // Set device path
-                details.device_path =
-                    Some(config_path.parent().unwrap().to_string_lossy().to_string());
-            } else {
-                log::debug!("Config file not found for {avd_name}: {config_path:?}");
-                // Set device path even if config file doesn't exist
-                let avd_path = std::path::PathBuf::from(&home_dir)
-                    .join(files::android::AVD_DIR)
-                    .join(files::android::AVD_SUBDIR)
-                    .join(format!("{avd_name}.avd"));
-                details.device_path = Some(avd_path.to_string_lossy().to_string());
-
-                // Set default values when config is not found
-                if details.ram_size.is_none() {
-                    details.ram_size = Some(format!("{} MB", defaults::DEFAULT_RAM_MB));
-                }
-                if details.storage_size.is_none() {
-                    details.storage_size = Some(format!("{} MB", defaults::DEFAULT_STORAGE_MB));
-                }
-                // Set default device type and API level if not found
-                if details.device_type.is_empty() {
-                    details.device_type = "Unknown Device".to_string();
-                }
-                if details.api_level_or_version.is_empty() {
-                    details.api_level_or_version = "Unknown Version".to_string();
-                }
-            }
-        } else {
-            log::warn!("HOME environment variable not set, cannot determine device path");
-        }
-
-        // Clean up resolution if we only got partial info
-        if let Some(ref res) = details.resolution {
-            if res.contains("?") {
-                details.resolution = None;
-            }
-        }
-
-        Ok(details)
     }
 
     /// Get appropriate skin name for device type using dynamic lookup
